add preemptive scheduling logic

+++ internal/controllers/scheduler/controller.go
@@ -70,6 +70,13 @@ func (r *ClusterScheduler) Reconcile(ctx context.Context, req reconcile.Request)
 			}
 			return clustersv1alpha1.REQUEST_GRANTED, nil
 		}).
+		WithCustomUpdateFunc(func(obj *clustersv1alpha1.ClusterRequest, rr ctrlutils.ReconcileResult[*clustersv1alpha1.ClusterRequest, clustersv1alpha1.ConditionStatus]) error {
+			// make sure to never write a cluster reference into the status for preemptive requests
+			if rr.Object != nil && rr.Object.Spec.Preemptive {
+				rr.Object.Status.Cluster = nil
+			}
+			return nil
+		}).
 		Build().
 		UpdateStatus(ctx, r.PlatformCluster.Client(), rr)
 }
@@ -171,54 +178,12 @@ func (r *ClusterScheduler) handleCreateOrUpdate(ctx context.Context, req reconci
 	}
 
 	// if no cluster was found, check if there is an existing cluster that qualifies for the request
-	if cDef.Template.Spec.Tenancy == clustersv1alpha1.TENANCY_SHARED {
-		log.Debug("Cluster template allows sharing, checking for fitting clusters", "purpose", purpose, "tenancyCount", cDef.TenancyCount)
-		// remove all clusters with a non-zero deletion timestamp from the list of candidates
-		clusters = filters.FilterSlice(clusters, func(args ...any) bool {
-			c, ok := args[0].(*clustersv1alpha1.Cluster)
-			if !ok {
-				return false
-			}
-			return c.DeletionTimestamp.IsZero()
-		})
-		// unless the cluster template for the requested purpose allows unlimited sharing, filter out all clusters that are already at their tenancy limit
-		if cDef.TenancyCount > 0 {
-			clusters = filters.FilterSlice(clusters, func(args ...any) bool {
-				c, ok := args[0].(*clustersv1alpha1.Cluster)
-				if !ok {
-					return false
-				}
-				return c.GetTenancyCount() < cDef.TenancyCount
-			})
-		}
-		if len(clusters) == 1 {
-			cluster = clusters[0]
-			log.Debug("One existing cluster qualifies for request", "clusterName", cluster.Name, "clusterNamespace", cluster.Namespace)
-		} else if len(clusters) > 0 {
-			log.Debug("Multiple existing clusters qualify for request, choosing one according to strategy", "strategy", string(r.Config.Strategy), "count", len(clusters))
-			switch r.Config.Strategy {
-			case config.STRATEGY_SIMPLE:
-				cluster = clusters[0]
-			case config.STRATEGY_RANDOM:
-				cluster = clusters[rand.IntN(len(clusters))]
-			case "":
-				// default to balanced, if empty
-				fallthrough
-			case config.STRATEGY_BALANCED:
-				// find cluster with least number of requests
-				cluster = clusters[0]
-				count := cluster.GetTenancyCount()
-				for _, c := range clusters[1:] {
-					tmp := c.GetTenancyCount()
-					if tmp < count {
-						count = tmp
-						cluster = c
-					}
-				}
-			default:
-				rr.ReconcileError = errutils.WithReason(fmt.Errorf("unknown strategy '%s'", r.Config.Strategy), cconst.ReasonConfigurationProblem)
-				return rr
-			}
+	// skip this check for preemptive requests with purposes with exclusive tenancy, as they will always result in a new cluster
+	if !(cDef.IsExclusive() && cr.Spec.Preemptive) {
+		cluster, rerr = r.pickFittingCluster(ctx, cr, clusters, cDef)
+		if rerr != nil {
+			rr.ReconcileError = rerr
+			return rr
 		}
 	}
 
@@ -228,8 +193,28 @@ func (r *ClusterScheduler) handleCreateOrUpdate(ctx context.Context, req reconci
 		// patch finalizer into Cluster
 		oldCluster := cluster.DeepCopy()
 		fin := cr.FinalizerForCluster()
-		if controllerutil.AddFinalizer(cluster, fin) {
-			log.Debug("Adding finalizer to cluster", "clusterName", cluster.Name, "clusterNamespace", cluster.Namespace, "finalizer", fin)
+		changed := controllerutil.AddFinalizer(cluster, fin) // should always be true at this point
+		replacedPreemptiveUID := ""
+		if !cr.Spec.Preemptive && !cDef.IsSharedUnlimitedly() {
+			// if the request is not preemptive, it can potentially replace a preemptive request that was already assigned to this cluster
+			// this check can be skipped for clusters which are shared unlimitedly, as the preemptive request would point to the same cluster again
+			suffix, removed := RemoveFinalizerWithPrefix(cluster, clustersv1alpha1.PreemptiveRequestFinalizerOnClusterPrefix)
+			replacedPreemptiveUID = suffix
+			if removed {
+				log.Info("Replacing preemptive request", "uid", replacedPreemptiveUID)
+				if rerr := r.ReschedulePreemptiveRequest(ctx, replacedPreemptiveUID); rerr != nil {
+					if rerr.Reason() == cconst.ReasonInternalError {
+						// if the problem is that the preemptive request was not found, only log an error
+						log.Error(rerr, "Error rescheduling preemptive request")
+					} else {
+						rr.ReconcileError = rerr
+						return rr
+					}
+				}
+			}
+		}
+		if changed {
+			log.Debug("Adding finalizer to cluster", "clusterName", cluster.Name, "clusterNamespace", cluster.Namespace, "finalizer", fin, "replacedPreemptiveUID", replacedPreemptiveUID)
 			if err := r.PlatformCluster.Client().Patch(ctx, cluster, client.MergeFrom(oldCluster)); err != nil {
 				rr.ReconcileError = errutils.WithReason(fmt.Errorf("error patching finalizer '%s' on cluster '%s/%s': %w", fin, cluster.Namespace, cluster.Name, err), cconst.ReasonPlatformClusterInteractionProblem)
 				return rr
@@ -295,7 +280,7 @@ func (r *ClusterScheduler) handleDelete(ctx context.Context, req reconcile.Reque
 				errs.Append(errutils.WithReason(fmt.Errorf("error patching finalizer '%s' on cluster '%s/%s': %w", fin, c.Namespace, c.Name, err), cconst.ReasonPlatformClusterInteractionProblem))
 			}
 		}
-		if c.GetTenancyCount() == 0 && ctrlutils.HasLabelWithValue(c, clustersv1alpha1.DeleteWithoutRequestsLabel, "true") {
+		if c.GetTenancyCount() == 0 && c.GetPreemptiveTenancyCount() == 0 && ctrlutils.HasLabelWithValue(c, clustersv1alpha1.DeleteWithoutRequestsLabel, "true") {
 			log.Info("Deleting cluster without requests", "clusterName", c.Name, "clusterNamespace", c.Namespace)
 			if err := r.PlatformCluster.Client().Delete(ctx, c); err != nil {
 				if apierrors.IsNotFound(err) {
@@ -343,6 +328,8 @@ func (r *ClusterScheduler) SetupWithManager(mgr ctrl.Manager) error {
 }
 
 // fetchRelevantClusters fetches all Cluster resources that could qualify for the given ClusterRequest.
+// These are all clusters in the expected namespace (for namespaced mode) that match the definition's label selector and have the requested purpose.
+// They are sorted by age, with the oldest cluster first. This reduces the chance of picking a cluster that is not ready yet because it was just created.
 func (r *ClusterScheduler) fetchRelevantClusters(ctx context.Context, cr *clustersv1alpha1.ClusterRequest, cDef *config.ClusterDefinition) ([]*clustersv1alpha1.Cluster, errutils.ReasonableError) {
 	// fetch clusters
 	purpose := cr.Spec.Purpose
@@ -372,6 +359,13 @@ func (r *ClusterScheduler) fetchRelevantClusters(ctx context.Context, cr *cluste
 		return slices.Contains(c.Spec.Purposes, purpose)
 	})
 
+	slices.SortStableFunc(clusters, func(a, b *clustersv1alpha1.Cluster) int {
+		if a == nil || b == nil {
+			return 0 // cannot compare nil clusters, should not happen
+		}
+		return a.CreationTimestamp.Time.Compare(b.CreationTimestamp.Time)
+	})
+
 	return clusters, nil
 }
 
@@ -389,7 +383,7 @@ func (r *ClusterScheduler) initializeNewCluster(ctx context.Context, cr *cluster
 	// - for exclusive clusters or shared limited:
 	//  1. generateName of template
 	//  2. purpose used as generateName
-	if cDef.Template.Spec.Tenancy == clustersv1alpha1.TENANCY_SHARED && cDef.TenancyCount == 0 {
+	if cDef.IsSharedUnlimitedly() {
 		// there will only be one instance of this cluster
 		if cDef.Template.GenerateName != "" {
 			cluster.SetGenerateName(cDef.Template.GenerateName)
@@ -415,7 +409,7 @@ func (r *ClusterScheduler) initializeNewCluster(ctx context.Context, cr *cluster
 	} else {
 		cluster.SetNamespace(cr.Namespace)
 	}
-	log.Info("Creating new cluster", "clusterName", cluster.Name, "clusterNamespace", cluster.Namespace)
+	log.Info("Creating new cluster", "clusterName", cluster.Name, "clusterNamespace", cluster.Namespace, "preemptive", cr.Spec.Preemptive)
 
 	// set finalizer
 	cluster.SetFinalizers([]string{cr.FinalizerForCluster()})
@@ -439,4 +433,127 @@ func (r *ClusterScheduler) initializeNewCluster(ctx context.Context, cr *cluster
 	}
 
 	return cluster
+}
+
+// TODO: move to controller-utils
+// RemoveFinalizerWithPrefix removes the first finalizer with the given prefix from the object.
+// The bool return value indicates whether a finalizer was removed.
+// If it is true, the string return value holds the suffix of the removed finalizer.
+// The logic is based on the controller-runtime's RemoveFinalizer function.
+func RemoveFinalizerWithPrefix(obj client.Object, prefix string) (string, bool) {
+	fins := obj.GetFinalizers()
+	length := len(fins)
+	suffix := ""
+
+	index := 0
+	for i := 0; i < length; i++ {
+		if strings.HasPrefix(fins[i], prefix) {
+			suffix = strings.TrimPrefix(fins[i], prefix)
+			continue
+		}
+		fins[index] = fins[i]
+		index++
+	}
+	obj.SetFinalizers(fins[:index])
+	return suffix, length != index
+}
+
+// ReschedulePreemptiveRequest fetches the ClusterRequest with the corresponding UID from the cluster and adds a reconcile annotation to it.
+func (r *ClusterScheduler) ReschedulePreemptiveRequest(ctx context.Context, uid string) errutils.ReasonableError {
+	crl := &clustersv1alpha1.ClusterRequestList{}
+	if err := r.PlatformCluster.Client().List(ctx, crl, client.MatchingLabelsSelector{Selector: r.Config.Selectors.Requests.Completed()}, client.MatchingFields{"spec.preemptive": "true"}); err != nil {
+		return errutils.WithReason(fmt.Errorf("error listing ClusterRequests: %w", err), cconst.ReasonPlatformClusterInteractionProblem)
+	}
+	var cr *clustersv1alpha1.ClusterRequest
+	for _, elem := range crl.Items {
+		if string(elem.UID) == uid {
+			cr = &elem
+			break
+		}
+	}
+	if cr == nil {
+		return errutils.WithReason(fmt.Errorf("unable to find preemptive ClusterRequest with UID '%s'", uid), cconst.ReasonInternalError)
+	}
+	if err := ctrlutils.EnsureAnnotation(ctx, r.PlatformCluster.Client(), cr, apiconst.OperationAnnotation, apiconst.OperationAnnotationValueReconcile, true); err != nil {
+		return errutils.WithReason(fmt.Errorf("error adding reconcile annotation to preemptive ClusterRequest '%s': %w", cr.Name, err), cconst.ReasonPlatformClusterInteractionProblem)
+	}
+	return nil
+}
+
+// pickFittingCluster gets a list of existing clusters that match the request's purpose and tries to pick one for the request.
+// If the returned cluster is nil, no fitting cluster was found. This can e.g. happen if all clusters are already at their tenancy limit.
+func (r *ClusterScheduler) pickFittingCluster(ctx context.Context, cr *clustersv1alpha1.ClusterRequest, clusters []*clustersv1alpha1.Cluster, cDef *config.ClusterDefinition) (*clustersv1alpha1.Cluster, errutils.ReasonableError) {
+	log := logging.FromContextOrPanic(ctx)
+	log.Debug("Checking for fitting clusters", "purpose", cr.Spec.Purpose, "tenancyCount", cDef.TenancyCount, "preemptive", cr.Spec.Preemptive)
+	// remove all clusters with a non-zero deletion timestamp from the list of candidates
+	clusters = filters.FilterSlice(clusters, func(args ...any) bool {
+		c, ok := args[0].(*clustersv1alpha1.Cluster)
+		if !ok {
+			return false
+		}
+		return c.DeletionTimestamp.IsZero()
+	})
+	// avoid picking clusters which are not ready if ready ones are available
+	readyClusters := filters.FilterSlice(clusters, func(args ...any) bool {
+		c, ok := args[0].(*clustersv1alpha1.Cluster)
+		if !ok {
+			return false
+		}
+		return c.Status.Phase == clustersv1alpha1.CLUSTER_PHASE_READY
+	})
+	if len(readyClusters) > 0 {
+		log.Debug("Excluding non-ready clusters from scheduling", "readyCount", len(readyClusters), "totalCount", len(clusters))
+		clusters = readyClusters
+	}
+	// unless the cluster template for the requested purpose allows unlimited sharing, filter out all clusters that are already at their tenancy limit
+	// for preemptive requests, other preemptive requests count towards the tenancy limit, for regular requests, preemptive requests are ignored
+	if cDef.IsExclusive() || cDef.TenancyCount > 0 {
+		clusters = filters.FilterSlice(clusters, func(args ...any) bool {
+			c, ok := args[0].(*clustersv1alpha1.Cluster)
+			if !ok {
+				return false
+			}
+			preTenCount := c.GetPreemptiveTenancyCount()
+			if cDef.IsExclusive() && preTenCount == 0 {
+				// for exclusive tenancy, take only clusters into account that have been preemptively requested
+				return false
+			}
+			tenCount := c.GetTenancyCount()
+			if cr.Spec.Preemptive {
+				// for preemptive requests, also take preemptive "workloads" into account
+				tenCount += preTenCount
+			}
+			return tenCount < cDef.TenancyCount
+		})
+	}
+	var cluster *clustersv1alpha1.Cluster
+	if len(clusters) == 1 {
+		cluster = clusters[0]
+		log.Debug("One existing cluster qualifies for request", "clusterName", cluster.Name, "clusterNamespace", cluster.Namespace, "preemptive", cr.Spec.Preemptive)
+	} else if len(clusters) > 0 {
+		log.Debug("Multiple existing clusters qualify for request, choosing one according to strategy", "strategy", string(r.Config.Strategy), "count", len(clusters), "preemptive", cr.Spec.Preemptive)
+		switch r.Config.Strategy {
+		case config.STRATEGY_SIMPLE:
+			cluster = clusters[0]
+		case config.STRATEGY_RANDOM:
+			cluster = clusters[rand.IntN(len(clusters))]
+		case "":
+			// default to balanced, if empty
+			fallthrough
+		case config.STRATEGY_BALANCED:
+			// find cluster with least number of requests
+			cluster = clusters[0]
+			count := cluster.GetTenancyCount()
+			for _, c := range clusters[1:] {
+				tmp := c.GetTenancyCount()
+				if tmp < count {
+					count = tmp
+					cluster = c
+				}
+			}
+		default:
+			return nil, errutils.WithReason(fmt.Errorf("unknown strategy '%s'", r.Config.Strategy), cconst.ReasonConfigurationProblem)
+		}
+	}
+	return cluster, nil
 }
